# backend/openapi.json
{
  "components": {
    "schemas": {
      "ApiResponse_Any_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "ApiResponse[Any]",
        "type": "object"
      },
...
      "AssistRequest": {
        "description": "写作辅助请求模型",
        "properties": {
          "action": {
            "default": "continue",
            "title": "Action",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "AssistRequest",
        "type": "object"
      },
      "AutocompleteRequest": {
        "description": "内联补全请求模型",
        "properties": {
          "context": {
            "title": "Context",
            "type": "string"
          }
        },
        "required": [
          "context"
        ],
        "title": "AutocompleteRequest",
        "type": "object"
      },
      "BatchCategoryRequest": {
        "description": "批量更新分类请求模型",
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "ids": {
            "items": {
              "type": "string"
            },
            "title": "Ids",
            "type": "array"
          }
        },
        "required": [
          "ids",
          "category"
        ],
        "title": "BatchCategoryRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/note/assist/stream": {
      "post": {
        "description": "AI 写作辅助 SSE 流式输出。支持三种模式：\n- continue：续写\n- expand：扩写\n- summarize：缩写",
        "operationId": "assist_stream_note_assist_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "example": "data: {\"schema_version\":\"1.0\",\"type\":\"response\",\"content\":\"hello\"}\n\n",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Versioned Server-Sent Events stream"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assist Stream",
        "tags": [
          "note"
        ]
      }
    },
    "/note/autocomplete": {
      "post": {
        "description": "AI 内联补全。基于光标前上下文，调用本地 Ollama qwen3:0.8b 快速返回续写文本。\n非流式，目标延迟 300-500ms。",
        "operationId": "autocomplete_note_autocomplete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutocompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Any_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Autocomplete",
        "tags": [
          "note"
        ]
      }
    },
    "/note/batch/category": {
      "put": {
        "description": "批量更新笔记分类。",
        "operationId": "batch_update_category_note_batch_category_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Batch Update Category",
        "tags": [
          "note"
        ]
      }
    },
    "/note/batch/delete": {
      "post": {
        "description": "批量删除笔记：按 ID 列表删除笔记及其向量。",
        "operationId": "batch_delete_notes_note_batch_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Batch Delete Notes",
        "tags": [
          "note"
        ]
      }
    },
    "/note/batch/download": {
      "post": {
        "description": "批量下载笔记为 ZIP 压缩包（内含 .md 文件）。",
        "operationId": "batch_download_notes_note_batch_download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Batch Download Notes",
        "tags": [
          "note"
        ]
      }
    },
    "/note/batch/pin": {
      "put": {
        "description": "批量置顶/取消置顶笔记。",
        "operationId": "batch_pin_notes_note_batch_pin_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Batch Pin Notes",
        "tags": [
          "note"
        ]
      }
    },
    "/note/category/{category}": {
      "delete": {
        "description": "删除某个分类及其下所有笔记。\n返回被删除的笔记数量。",
        "operationId": "delete_category_note_category__category__delete",
        "parameters": [
          {
            "in": "path",
            "name": "category",
            "required": true,
            "schema": {
              "title": "Category",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Any_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Category",
        "tags": [
          "note"
        ]
      }
    },
    "/note/create": {
      "post": {
        "description": "创建笔记：\n1. MySQL 写入 + ChromaDB 向量化\n2. 立即返回笔记（tags/category 初始为空）\n3. 后台异步生成标签和复习记忆事项",
        "operationId": "create_note_note_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoteResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Note",
        "tags": [
          "note"
        ]
      }
    },
    "/note/list": {
      "get": {
        "description": "笔记列表：分页查询，支持按分类筛选和排序。tag 筛选在内存层完成。",
        "operationId": "list_notes_note_list_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "title": "Category",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "title": "Tag",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "updated_at",
              "pattern": "^(updated_at|created_at|title)$",
              "title": "Sort By",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoteListResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Notes",
        "tags": [
          "note"
        ]
      }
    },
    "/note/search": {
      "get": {
        "description": "全文语义搜索：走 ChromaDB notes_collection 向量检索，\n返回当前用户的语义相似笔记。",
        "operationId": "search_notes_note_search_get",
        "parameters": [
          {
            "description": "搜索关键词",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "搜索关键词",
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoteListResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Notes",
        "tags": [
          "note"
        ]
      }
    },
    "/note/stats": {
      "get": {
        "description": "获取用户笔记分类统计。\n返回各分类下的笔记数量及总数。",
        "operationId": "get_stats_note_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Any_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Stats",
        "tags": [
          "note"
        ]
      }
    },
    "/note/{note_id}": {
      "delete": {
        "description": "删除笔记：联删 MySQL 记录、ChromaDB 向量、以及对应复习记忆事项。",
        "operationId": "delete_note_note__note_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Note",
        "tags": [
          "note"
        ]
      },
      "get": {
        "description": "获取笔记详情。",
        "operationId": "get_note_note__note_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoteResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Note",
        "tags": [
          "note"
        ]
      },
      "put": {
        "description": "更新笔记：修改 title/content，content 变更时同步更新 ChromaDB 向量。",
        "operationId": "update_note_note__note_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoteResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Note",
        "tags": [
          "note"
        ]
      }
    },
    "/note/{note_id}/auto-tag": {
      "post": {
        "description": "手动触发重新生成标签。",
        "operationId": "regenerate_tags_note__note_id__auto_tag_post",
        "parameters": [
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Regenerate Tags",
        "tags": [
          "note"
        ]
      }
    },
    "/note/{note_id}/download": {
      "get": {
        "description": "下载笔记为 Markdown 文件（浏览器触发下载）。\n返回 Content-Disposition: attachment 的 markdown 文件响应。",
        "operationId": "download_note_note__note_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              },
              "text/markdown": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Markdown download, or a canonical JSON envelope when absent"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Note",
        "tags": [
          "note"
        ]
      }
    },
    "/note/{note_id}/export": {
      "get": {
        "description": "导出单篇笔记为 Markdown 格式纯文本。",
        "operationId": "export_note_note__note_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Any_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Note",
        "tags": [
          "note"
        ]
      }
    },
...
  }
}